# Week-5/07-API-Documentation-using-OpenAPI-Spec/07-Rest-APIs-with-Swagger-Docs/uploads/openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "YT Node js APIs",
      "description": "This is Swagger Documetation",
      "version": "1.0.0",
      "contact": {}
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
    "paths": {
      "/api/v1/auth/register": {
        "post": {
          "summary": "Register User API",
          "description": "",
          "operationId": "postapiv1authregister",
          "parameters": [],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "example": {}
                  }
                }
              }
            }
          },
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "name": "User 8",
                    "email": "user8@gmail.com",
                    "password": "12345678",
                    "gender": "male"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/auth/login": {
        "post": {
          "summary": "Login User API",
          "description": "",
          "operationId": "postapiv1authlogin",
          "parameters": [],
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "example": {}
                  }
                }
              }
            }
          },
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "email": "user8@gmail.com",
                    "password": "12345678"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
...
      "securitySchemes": {
        "basicAuth": {
          "type": "http",
          "scheme": "basic"
        },
        "digestAuth": {
          "type": "http",
          "scheme": "digest"
        }
      }
    },
...
  }